Note that the plugin is deprecated
https://github.com/shillner/unleash-maven-plugin/ has been archived.
That is the unleash plugin for Apache Maven.  The Jenkins unleash maven
plugin depends on the unleash plugin for Apache Maven.

https://github.com/jenkinsci/unleash-plugin has been archived.  That is
the Jenkins unleash Maven plugin. The last release of the Jenkins unleash
maven plugin was 3 years ago. Since the Jenkins plugin repository has
been archived, we can expect no new releases of the plugin.

[JENKINS-70326](https://issues.jenkins.io/browse/JENKINS-70326) notes
that the Jenkins unleash maven plugin does not work with Java 17.

Let's deprecate the Jenkins plugin so that it is clear to users that the
Jenkins plugin is not being maintained and that the Apache Maven plugin
is not being maintained.

jenkins-infra/repository-permissions-updater#3042
is the request to remove all developers from the archived plugin
repository